# Optimized BuildHist function