Optimizations for Histogram building. A part of the issue #5104
The PR contains changes from #5138, it will be rebased after #5138 merging.