@SmirnovEgorRu 2019-12-31
Current performance using previous commit https://github.com/dmlc/xgboost/pull/5138 also:

higgs1m | ApplySplit | EvaluateSplit | BuildHist | SyncHistogram | Prediction | Total, sec
-- | -- | -- | -- | -- | -- | --
Master | 33 | 29 | 90 | 26 | 3 | 185
Before   reverting | 3.7 | 3.5 | 6.2 | 0.0 | 1.6 | 17.7
This PR | 27.7 | 1.7 | 9.5 | 2.1 | 1.6 | 47.3

airline-ohe | ApplySplit | EvaluateSplit | BuildHist | SyncHistogram | Prediction | Total, sec
-- | -- | -- | -- | -- | -- | --
Master | 26 | 27 | 67 | 12 | 2 | 157
Before   reverting | 9.0 | 6.1 | 28.8 | 0.0 | 0.7 | 63.7
This PR | 21.4 | 2.9 | 42.2 | 1.0 | 0.7 | 93.5