@SmirnovEgorRu 2019-12-31
@hcho3, @trivialfis, I finalized the PR from my side. Could you, please, look at this?
It contains changes from #5138, I will do rebase after merging previous one into master.