@hcho3 2020-01-08
One distributed test is stuck: https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/PR-5156/8/pipeline/112. I had to kill it by hand. I'm looking at the code now to see what went wrong; probably a worker is not calling AllReduce().