@SmirnovEgorRu 2020-01-09
@hcho3, I have fixed the issue.
I also tested performance of distributed mode, for small data sets - it is similar as before, for large - gain is observed. It's reasonable, because # of "AllReduce" calls are the same.

CC @trivialfis