@hcho3 2020-01-16
Reminder to myself: Write a follow-up PR so that we can control how many threads `common::ParallelFor2d` will launch.