@trivialfis 2020-01-16
Yup.  Will review tonight.  Sorry for the long wait.