@trivialfis 2020-01-16
@hcho3 

> Reminder to myself: Write a follow-up PR so that we can control how many threads common::ParallelFor2d will launch.

I'm a little bit concern that, it's possible for a user to change `nthread` during training.