@hcho3 2020-01-17
@trivialfis I think `nthread` is a configurable parameter. Once it's set, I don't think it would change in the middle of training.