@trivialfis 2020-01-17
@SmirnovEgorRu  Could you please take a look into this dataset:  http://archive.ics.uci.edu/ml/machine-learning-databases/url/url_svmlight.tar.gz ?  It's extremely sparse, my benchmarks shows regression on both training time and memory usage:

Before:
```
System  : 3.7.3 (default, Apr  3 2019, 05:39:12)
[GCC 8.3.0]
Xgboost : 1.0.0-SNAPSHOT
LightGBM: None
CatBoost: None
#jobs   : 6
Running 'xgb-cpu' ...
{
  "url": {
    "xgb-cpu": {
      "accuracy": {
        "AUC": 0.9920843803290437,
        "Accuracy": 0.9776493762859277,
        "Log_Loss": 0.4229395458885813,
        "Precision": 0.9781071399946979,
        "Recall": 0.9887536809959334
      },
      "train_time": 172.10726167000007
    }
  }
}
Results written to file 'url.json'
Child terminated but following descendants are still running: 4483

Could not remove sub-cgroup /sys/fs/cgroup/memory/cgmemtime/4462: Device or resource busy
Child user:  400.110 s
Child sys :   70.966 s
Child wall:  229.677 s
Child high-water RSS                    :   34821560 KiB
Recursive and acc. high-water RSS+CACHE :   38068864 KiB
```

After:
```
System  : 3.7.3 (default, Apr  3 2019, 05:39:12)
[GCC 8.3.0]
Xgboost : 1.0.0-SNAPSHOT
LightGBM: None
CatBoost: None
#jobs   : 6
Running 'xgb-cpu' ...
{
  "url": {
    "xgb-cpu": {
      "accuracy": {
        "AUC": 0.9920843803290437,
        "Accuracy": 0.9776493762859277,
        "Log_Loss": 0.4229395458885813,
        "Precision": 0.9781071399946979,
        "Recall": 0.9887536809959334
      },
      "train_time": 203.096213423
    }
  }
}
Results written to file 'url.json'
Child terminated but following descendants are still running: 3351

Could not remove sub-cgroup /sys/fs/cgroup/memory/cgmemtime/3330: Device or resource busy
Child user:  408.533 s
Child sys :   82.425 s
Child wall:  263.838 s
Child high-water RSS                    :   46879656 KiB
Recursive and acc. high-water RSS+CACHE :   46983308 KiB
```