@trivialfis 2020-01-21
@SmirnovEgorRu BTW, the memory usage is measured by https://github.com/gsauthof/cgmemtime .