@SmirnovEgorRu 2020-01-26
@trivialfis @hcho3 
Thank you for your tests, I added additional changes to reduce amount of partial histograms. As result for URL data set I obtained following numbers:

|Branch|Time of Update, sec|Memory usage, KB|
|--|--|--|
|master|74.51|24638840 |
|this PR|23.53|24109408|

So, now it's better for both execution time and memory consumption.