@SmirnovEgorRu 2020-01-26
@hcho3, @trivialfis, CI is also green. Do you see any road blockers to merge the pull-request?