@SmirnovEgorRu 2020-01-29
@hcho3 @trivialfis, I have already created new PR https://github.com/dmlc/xgboost/pull/5244 which should finalize efforts on reverting the optimizations. Could you, please, accept or provide new comments for current PR to enable review of the next one?