@SmirnovEgorRu 2020-01-29
@trivialfis, let me understand. Do we plan to include the change to 1.0 version as it was discussed originally https://github.com/dmlc/xgboost/pull/5008#issuecomment-562386267 with @hcho3 ?