@trivialfis 2020-01-29
I don't plan to add major changes at the last minute.  It's fine as we used to have a release every 2 or 3 months, and I would like to resume the pace once this 1.0 thing is finished.  So your changes should be available in short term even not in the next release.  Besides there's nightly build available for download.

Will merge once we have a release branch.