@hcho3 2020-01-29
@trivialfis I agree that, in general, we should not merge a major change right before a release. However, since you and I have approved the original version of this PR, can we merge this?