@SmirnovEgorRu 2020-01-29
@hcho3 @trivialfis, if you need some specific testing like running specified benchmarks or workloads to have more confidence - I'm happy to help here.

I am grateful for the chance to have this in the nearest XGB release