@trivialfis 2020-01-29
@SmirnovEgorRu Basically memory usage, computation time, accuracy (auc, rmse metrics) for these representative datasets (like higgs for dense wide columns, url for sparse), with restricted number of threads, max_depth, num_boosted_rounds etc,  and maybe manually set cpu affinity env for OMP.  I normally do this myself as I can have a consistent environment for each run.  For example the number posted in:

https://github.com/dmlc/xgboost/pull/5156#issuecomment-578535021

and 

https://github.com/dmlc/xgboost/pull/5244#issuecomment-579867020

Seems to be running on different environment or with different parameters.