@hcho3 2020-01-29
@trivialfis I am confident about this PR. I’m inclined to merge, as long as @SmirnovEgorRu runs some more benchmarks as requested.