@trivialfis 2020-01-29
@hcho3 Got it.  My concerns are mostly around the consistency of posted number across different PRs.  As noted above, it would be nice to have benchmarks performed on same platform with fixed parameters.  The variance is sometimes difficult to control.