@SmirnovEgorRu 2020-01-30
@hcho3 @trivialfis , I prepared measurements on Higgs and Airline.

**Higgs:**

nthreads | 1 | 8 | 24 | 48 | 96
-- | -- | -- | -- | -- | --
This PR, sec | 280.9 | 87.4 | 48.4 | 44.8 | 45.3
master, sec | 281.6 | 145.9 | 135.6 | 167.6 | 376.5

Log-loss in all cases this PR/master:
LogLoss for train data set = 0.404697
LogLoss for test  data set = 0.525144

niter | 50 | 200 | 500 | 1000
-- | -- | -- | -- | --
This PR, sec | 3.5 | 10.7 | 22.7 | 44.8
master, sec | 13.7 | 41.4 | 91.8 | 167.6

**Airline + one-hot-encoding:**

nthreads | 1 | 8 | 24 | 48 | 96
-- | -- | -- | -- | -- | --
This PR, sec | 953.7 | 211.4 | 105 | 94.65 | 87.09
master, sec | 949.3 | 265.5 | 159.6 | 142.8 | 272.5

Log-loss in all cases this PR/master:
LogLoss for train data set = 0.383229
LogLoss for test  data set = 0.461403

niter | 50 | 200 | 500 | 1000
-- | -- | -- | -- | --
This PR, sec | 28.82 | 40.03 | 60.41 | 94.65
master, sec | 35.65 | 57.2 | 89.37 | 142.8


HW: AWS c5.metal, CLX @3.0GHz, 24 cores per socket, 2 sockets, HT: on, 96 threads totally